requirements.txt
fastapi[standard]
sqlmodel
pydantic-settings
alembic
passlib
bcrypt
pyjwt
redis